% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{conv_age_vars}
\alias{conv_age_vars}
\title{Convert e-Stat's age labels}
\usage{
conv_age_vars(x)
}
\arguments{
\item{x}{variable}
}
\description{
Normalize age labels that differ between Population Census tables.
}
