% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{conv_gender_vars}
\alias{conv_gender_vars}
\title{Convert e-Stat's record value}
\usage{
conv_gender_vars(x, lang = "ja")
}
\arguments{
\item{x}{variable}

\item{lang}{Specify the language to convert}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
