% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpops-package.R
\docType{package}
\name{jpops-package}
\alias{jpops}
\alias{jpops-package}
\title{jpops: Provides a Data Set on the Population in Japan}
\description{
Information on the population and household registers, as published by the census and each local authority. The acquired data will be returned in a tidy format.
}
\author{
\strong{Maintainer}: Shinya Uryu \email{suika1127@gmail.com} (\href{https://orcid.org/0000-0002-0493-6186}{ORCID})

}
\keyword{internal}
